xds: refactor the ads_mux from the cluster-manager to the xDS-manager
