+++ envoy/upstream/cluster_manager.h
@@ -394,15 +394,6 @@ public:
   virtual Config::GrpcMuxSharedPtr adsMux() PURE;
 
   /**
-   * Replaces the current ADS mux with a new one based on the given config.
-   * Assumes that the given ads_config is syntactically valid (according to the PGV constraints).
-   * @param ads_config an ADS config source to use.
-   * @return the status of the operation.
-   */
-  virtual absl::Status
-  replaceAdsMux(const envoy::config::core::v3::ApiConfigSource& ads_config) PURE;
-
-  /**
    * @return Grpc::AsyncClientManager& the cluster manager's gRPC client manager.
    */
   virtual Grpc::AsyncClientManager& grpcAsyncClientManager() PURE;
