+++ source/common/upstream/cluster_manager_impl.cc
@@ -24,7 +24,6 @@
 #include "source/common/common/enum_to_int.h"
 #include "source/common/common/fmt.h"
 #include "source/common/common/utility.h"
-#include "source/common/config/custom_config_validators_impl.h"
 #include "source/common/config/null_grpc_mux_impl.h"
 #include "source/common/config/utility.h"
 #include "source/common/config/xds_resource.h"
@@ -105,21 +104,6 @@ bool isBlockingAdsCluster(const envoy::config::bootstrap::v3::Bootstrap& bootstr
     }
   }
   return blocking_ads_cluster;
-}
-
-absl::Status createClients(Grpc::AsyncClientFactoryPtr& primary_factory,
-                           Grpc::AsyncClientFactoryPtr& failover_factory,
-                           Grpc::RawAsyncClientPtr& primary_client,
-                           Grpc::RawAsyncClientPtr& failover_client) {
-  absl::StatusOr<Grpc::RawAsyncClientPtr> success = primary_factory->createUncachedRawAsyncClient();
-  RETURN_IF_NOT_OK_REF(success.status());
-  primary_client = std::move(*success);
-  if (failover_factory) {
-    success = failover_factory->createUncachedRawAsyncClient();
-    RETURN_IF_NOT_OK_REF(success.status());
-    failover_client = std::move(*success);
-  }
-  return absl::OkStatus();
 }
 
 } // namespace
@@ -322,10 +306,9 @@ ClusterManagerImpl::ClusterManagerImpl(
     Server::Configuration::CommonFactoryContext& context, Stats::Store& stats,
     ThreadLocal::Instance& tls, Runtime::Loader& runtime, const LocalInfo::LocalInfo& local_info,
     AccessLog::AccessLogManager& log_manager, Event::Dispatcher& main_thread_dispatcher,
-    OptRef<Server::Admin> admin, ProtobufMessage::ValidationContext& validation_context,
-    Api::Api& api, Http::Context& http_context, Grpc::Context& grpc_context,
-    Router::Context& router_context, Server::Instance& server, Config::XdsManager& xds_manager,
-    absl::Status& creation_status)
+    OptRef<Server::Admin> admin, Api::Api& api, Http::Context& http_context,
+    Grpc::Context& grpc_context, Router::Context& router_context, Server::Instance& server,
+    Config::XdsManager& xds_manager, absl::Status& creation_status)
     : server_(server), factory_(factory), runtime_(runtime), stats_(stats), tls_(tls),
       xds_manager_(xds_manager), random_(api.randomGenerator()),
       deferred_cluster_creation_(bootstrap.cluster_manager().enable_deferred_cluster_creation()),
@@ -336,8 +319,8 @@ ClusterManagerImpl::ClusterManagerImpl(
       init_helper_(*this,
                    [this](ClusterManagerCluster& cluster) { return onClusterInit(cluster); }),
       time_source_(main_thread_dispatcher.timeSource()), dispatcher_(main_thread_dispatcher),
-      http_context_(http_context), validation_context_(validation_context),
-      router_context_(router_context), cluster_stat_names_(stats.symbolTable()),
+      http_context_(http_context), router_context_(router_context),
+      cluster_stat_names_(stats.symbolTable()),
       cluster_config_update_stat_names_(stats.symbolTable()),
       cluster_lb_stat_names_(stats.symbolTable()),
       cluster_endpoint_stat_names_(stats.symbolTable()),
@@ -422,98 +405,8 @@ ClusterManagerImpl::initialize(const envoy::config::bootstrap::v3::Bootstrap& bo
     }
   }
 
-  const auto& dyn_resources = bootstrap.dynamic_resources();
-
   // Now setup ADS if needed, this might rely on a primary cluster.
-  // This is the only point where distinction between delta ADS and state-of-the-world ADS is made.
-  // After here, we just have a GrpcMux interface held in ads_mux_, which hides
-  // whether the backing implementation is delta or SotW.
-  if (dyn_resources.has_ads_config()) {
-    Config::CustomConfigValidatorsPtr custom_config_validators =
-        std::make_unique<Config::CustomConfigValidatorsImpl>(
-            validation_context_.dynamicValidationVisitor(), server_,
-            dyn_resources.ads_config().config_validators());
-
-    auto strategy_or_error = Config::Utility::prepareJitteredExponentialBackOffStrategy(
-        dyn_resources.ads_config(), random_,
-        Envoy::Config::SubscriptionFactory::RetryInitialDelayMs,
-        Envoy::Config::SubscriptionFactory::RetryMaxDelayMs);
-    RETURN_IF_NOT_OK_REF(strategy_or_error.status());
-    JitteredExponentialBackOffStrategyPtr backoff_strategy = std::move(strategy_or_error.value());
-
-    const bool use_eds_cache =
-        Runtime::runtimeFeatureEnabled("envoy.restart_features.use_eds_cache_for_ads");
-
-    if (dyn_resources.ads_config().api_type() ==
-        envoy::config::core::v3::ApiConfigSource::DELTA_GRPC) {
-      absl::Status status = Config::Utility::checkTransportVersion(dyn_resources.ads_config());
-      RETURN_IF_NOT_OK(status);
-      std::string name;
-      if (Runtime::runtimeFeatureEnabled("envoy.reloadable_features.unified_mux")) {
-        name = "envoy.config_mux.delta_grpc_mux_factory";
-      } else {
-        name = "envoy.config_mux.new_grpc_mux_factory";
-      }
-      auto* factory = Config::Utility::getFactoryByName<Config::MuxFactory>(name);
-      if (!factory) {
-        return absl::InvalidArgumentError(fmt::format("{} not found", name));
-      }
-      auto factory_primary_or_error = Config::Utility::factoryForGrpcApiConfigSource(
-          *async_client_manager_, dyn_resources.ads_config(), *stats_.rootScope(), false, 0);
-      RETURN_IF_NOT_OK_REF(factory_primary_or_error.status());
-      Grpc::AsyncClientFactoryPtr factory_failover = nullptr;
-      if (Runtime::runtimeFeatureEnabled("envoy.restart_features.xds_failover_support")) {
-        auto factory_failover_or_error = Config::Utility::factoryForGrpcApiConfigSource(
-            *async_client_manager_, dyn_resources.ads_config(), *stats_.rootScope(), false, 1);
-        RETURN_IF_NOT_OK_REF(factory_failover_or_error.status());
-        factory_failover = std::move(factory_failover_or_error.value());
-      }
-      Grpc::RawAsyncClientPtr primary_client;
-      Grpc::RawAsyncClientPtr failover_client;
-      RETURN_IF_NOT_OK(createClients(factory_primary_or_error.value(), factory_failover,
-                                     primary_client, failover_client));
-      ads_mux_ =
-          factory->create(std::move(primary_client), std::move(failover_client), dispatcher_,
-                          random_, *stats_.rootScope(), dyn_resources.ads_config(), local_info_,
-                          std::move(custom_config_validators), std::move(backoff_strategy),
-                          xds_manager_.xdsConfigTracker(), {}, use_eds_cache);
-    } else {
-      absl::Status status = Config::Utility::checkTransportVersion(dyn_resources.ads_config());
-      RETURN_IF_NOT_OK(status);
-      std::string name;
-      if (Runtime::runtimeFeatureEnabled("envoy.reloadable_features.unified_mux")) {
-        name = "envoy.config_mux.sotw_grpc_mux_factory";
-      } else {
-        name = "envoy.config_mux.grpc_mux_factory";
-      }
-
-      auto* factory = Config::Utility::getFactoryByName<Config::MuxFactory>(name);
-      if (!factory) {
-        return absl::InvalidArgumentError(fmt::format("{} not found", name));
-      }
-      auto factory_primary_or_error = Config::Utility::factoryForGrpcApiConfigSource(
-          *async_client_manager_, dyn_resources.ads_config(), *stats_.rootScope(), false, 0);
-      RETURN_IF_NOT_OK_REF(factory_primary_or_error.status());
-      Grpc::AsyncClientFactoryPtr factory_failover = nullptr;
-      if (Runtime::runtimeFeatureEnabled("envoy.restart_features.xds_failover_support")) {
-        auto factory_failover_or_error = Config::Utility::factoryForGrpcApiConfigSource(
-            *async_client_manager_, dyn_resources.ads_config(), *stats_.rootScope(), false, 1);
-        RETURN_IF_NOT_OK_REF(factory_failover_or_error.status());
-        factory_failover = std::move(factory_failover_or_error.value());
-      }
-      Grpc::RawAsyncClientPtr primary_client;
-      Grpc::RawAsyncClientPtr failover_client;
-      RETURN_IF_NOT_OK(createClients(factory_primary_or_error.value(), factory_failover,
-                                     primary_client, failover_client));
-      ads_mux_ = factory->create(std::move(primary_client), std::move(failover_client), dispatcher_,
-                                 random_, *stats_.rootScope(), dyn_resources.ads_config(),
-                                 local_info_, std::move(custom_config_validators),
-                                 std::move(backoff_strategy), xds_manager_.xdsConfigTracker(),
-                                 xds_manager_.xdsResourcesDelegate(), use_eds_cache);
-    }
-  } else {
-    ads_mux_ = std::make_unique<Config::NullGrpcMuxImpl>();
-  }
+  RETURN_IF_NOT_OK(xds_manager_.initializeAdsConnections(bootstrap));
 
   // After ADS is initialized, load EDS static clusters as EDS config may potentially need ADS.
   for (const auto& cluster : bootstrap.static_resources().clusters()) {
@@ -555,6 +448,7 @@ ClusterManagerImpl::initialize(const envoy::config::bootstrap::v3::Bootstrap& bo
     return std::make_shared<ThreadLocalClusterManagerImpl>(*this, dispatcher, local_cluster_params);
   });
 
+  const auto& dyn_resources = bootstrap.dynamic_resources();
   // We can now potentially create the CDS API once the backing cluster exists.
   if (dyn_resources.has_cds_config() || !dyn_resources.cds_resources_locator().empty()) {
     std::unique_ptr<xds::core::v3::ResourceLocator> cds_resources_locator;
@@ -588,85 +482,9 @@ ClusterManagerImpl::initialize(const envoy::config::bootstrap::v3::Bootstrap& bo
   if (!has_ads_cluster) {
     // There is no ADS cluster, so we won't be starting the ADS mux after a cluster has finished
     // initializing, so we must start ADS here.
-    ads_mux_->start();
+    xds_manager_.adsMux()->start();
   }
   return absl::OkStatus();
-}
-
-absl::Status
-ClusterManagerImpl::replaceAdsMux(const envoy::config::core::v3::ApiConfigSource& ads_config) {
-  // If there was no ADS defined, reject replacement.
-  const auto& bootstrap = server_.bootstrap();
-  if (!bootstrap.has_dynamic_resources() || !bootstrap.dynamic_resources().has_ads_config()) {
-    return absl::InternalError(
-        "Cannot replace an ADS config when one wasn't previously configured in the bootstrap");
-  }
-  const auto& bootstrap_ads_config = server_.bootstrap().dynamic_resources().ads_config();
-
-  // There is no support for switching between different ADS types.
-  if (ads_config.api_type() != bootstrap_ads_config.api_type()) {
-    return absl::InternalError(fmt::format(
-        "Cannot replace an ADS config with a different api_type (expected: {})",
-        envoy::config::core::v3::ApiConfigSource::ApiType_Name(bootstrap_ads_config.api_type())));
-  }
-
-  // There is no support for using a different config validator. Note that if
-  // this is mainly because the validator could be stateful and if the delta-xDS
-  // protocol is used, then the new validator will not have the context of the
-  // previous one.
-  if (bootstrap_ads_config.config_validators_size() != ads_config.config_validators_size()) {
-    return absl::InternalError(fmt::format(
-        "Cannot replace config_validators in ADS config (different size) - Previous: {}, New: {}",
-        bootstrap_ads_config.config_validators_size(), ads_config.config_validators_size()));
-  } else if (bootstrap_ads_config.config_validators_size() > 0) {
-    const bool equal_config_validators = std::equal(
-        bootstrap_ads_config.config_validators().begin(),
-        bootstrap_ads_config.config_validators().end(), ads_config.config_validators().begin(),
-        [](const envoy::config::core::v3::TypedExtensionConfig& a,
-           const envoy::config::core::v3::TypedExtensionConfig& b) {
-          return Protobuf::util::MessageDifferencer::Equivalent(a, b);
-        });
-    if (!equal_config_validators) {
-      return absl::InternalError(fmt::format("Cannot replace config_validators in ADS config "
-                                             "(different contents)\nPrevious: {}\nNew: {}",
-                                             bootstrap_ads_config.DebugString(),
-                                             ads_config.DebugString()));
-    }
-  }
-
-  ENVOY_LOG_MISC(trace, "Replacing ADS config with:\n{}", ads_config.DebugString());
-  auto strategy_or_error = Config::Utility::prepareJitteredExponentialBackOffStrategy(
-      ads_config, random_, Envoy::Config::SubscriptionFactory::RetryInitialDelayMs,
-      Envoy::Config::SubscriptionFactory::RetryMaxDelayMs);
-  RETURN_IF_NOT_OK_REF(strategy_or_error.status());
-  JitteredExponentialBackOffStrategyPtr backoff_strategy = std::move(strategy_or_error.value());
-
-  absl::Status status = Config::Utility::checkTransportVersion(ads_config);
-  RETURN_IF_NOT_OK(status);
-
-  auto factory_primary_or_error = Config::Utility::factoryForGrpcApiConfigSource(
-      *async_client_manager_, ads_config, *stats_.rootScope(), false, 0);
-  RETURN_IF_NOT_OK_REF(factory_primary_or_error.status());
-  Grpc::AsyncClientFactoryPtr factory_failover = nullptr;
-  if (Runtime::runtimeFeatureEnabled("envoy.restart_features.xds_failover_support")) {
-    auto factory_failover_or_error = Config::Utility::factoryForGrpcApiConfigSource(
-        *async_client_manager_, ads_config, *stats_.rootScope(), false, 1);
-    RETURN_IF_NOT_OK_REF(factory_failover_or_error.status());
-    factory_failover = std::move(factory_failover_or_error.value());
-  }
-  Grpc::RawAsyncClientPtr primary_client;
-  Grpc::RawAsyncClientPtr failover_client;
-  RETURN_IF_NOT_OK(createClients(factory_primary_or_error.value(), factory_failover, primary_client,
-                                 failover_client));
-
-  // Primary client must not be null, as the primary xDS source must be a valid one.
-  // The failover_client may be null (no failover defined).
-  ASSERT(primary_client != nullptr);
-
-  // This will cause a disconnect from the current sources, and replacement of the clients.
-  status = ads_mux_->updateMuxSource(std::move(primary_client), std::move(failover_client),
-                                     *stats_.rootScope(), std::move(backoff_strategy), ads_config);
-  return status;
 }
 
 absl::Status ClusterManagerImpl::initializeSecondaryClusters(
@@ -1128,10 +946,10 @@ void ClusterManagerImpl::updateClusterCounts() {
   // If we're in the middle of shutting down (ads_mux_ already gone) then this is irrelevant.
   const bool all_clusters_initialized =
       init_helper_.state() == ClusterManagerInitHelper::State::AllClustersInitialized;
-  if (all_clusters_initialized && ads_mux_) {
+  if (all_clusters_initialized && xds_manager_.adsMux()) {
     const auto type_url = Config::getTypeUrl<envoy::config::cluster::v3::Cluster>();
     if (resume_cds_ == nullptr && !warming_clusters_.empty()) {
-      resume_cds_ = ads_mux_->pause(type_url);
+      resume_cds_ = xds_manager_.adsMux()->pause(type_url);
     } else if (warming_clusters_.empty()) {
       resume_cds_.reset();
     }
@@ -1444,7 +1262,7 @@ void ClusterManagerImpl::postThreadLocalClusterUpdate(ClusterManagerCluster& cm_
   // By this time, the main thread has received the cluster initialization update, so we can start
   // the ADS mux if the ADS mux is dependent on this cluster's initialization.
   if (cm_cluster.requiredForAds() && !ads_mux_initialized_) {
-    ads_mux_->start();
+    xds_manager_.adsMux()->start();
     ads_mux_initialized_ = true;
   }
 }
@@ -1822,8 +1640,8 @@ void ClusterManagerImpl::notifyClusterDiscoveryStatus(absl::string_view name,
 
 Config::EdsResourcesCacheOptRef ClusterManagerImpl::edsResourcesCache() {
   // EDS caching is only supported for ADS.
-  if (ads_mux_) {
-    return ads_mux_->edsResourcesCache();
+  if (xds_manager_.adsMux()) {
+    return xds_manager_.adsMux()->edsResourcesCache();
   }
   return {};
 }
@@ -2326,8 +2144,8 @@ absl::StatusOr<ClusterManagerPtr> ProdClusterManagerFactory::clusterManagerFromP
   auto cluster_manager_impl = std::unique_ptr<ClusterManagerImpl>{new ClusterManagerImpl(
       bootstrap, *this, context_, stats_, tls_, context_.runtime(), context_.localInfo(),
       context_.accessLogManager(), context_.mainThreadDispatcher(), context_.admin(),
-      context_.messageValidationContext(), context_.api(), http_context_, context_.grpcContext(),
-      context_.routerContext(), server_, context_.xdsManager(), creation_status)};
+      context_.api(), http_context_, context_.grpcContext(), context_.routerContext(), server_,
+      context_.xdsManager(), creation_status)};
   RETURN_IF_NOT_OK(creation_status);
   return cluster_manager_impl;
 }
