+++ source/server/config_validation/cluster_manager.cc
@@ -14,8 +14,8 @@ absl::StatusOr<ClusterManagerPtr> ValidationClusterManagerFactory::clusterManage
   auto cluster_manager = std::unique_ptr<ValidationClusterManager>{new ValidationClusterManager(
       bootstrap, *this, context_, stats_, tls_, context_.runtime(), context_.localInfo(),
       context_.accessLogManager(), context_.mainThreadDispatcher(), context_.admin(),
-      context_.messageValidationContext(), context_.api(), http_context_, context_.grpcContext(),
-      context_.routerContext(), server_, context_.xdsManager(), creation_status)};
+      context_.api(), http_context_, context_.grpcContext(), context_.routerContext(), server_,
+      context_.xdsManager(), creation_status)};
   RETURN_IF_NOT_OK(creation_status);
   return cluster_manager;
 }
