+++ test/common/upstream/test_cluster_manager.h
@@ -172,14 +172,13 @@ public:
       Server::Configuration::CommonFactoryContext& context, Stats::Store& stats,
       ThreadLocal::Instance& tls, Runtime::Loader& runtime, const LocalInfo::LocalInfo& local_info,
       AccessLog::AccessLogManager& log_manager, Event::Dispatcher& main_thread_dispatcher,
-      Server::Admin& admin, ProtobufMessage::ValidationContext& validation_context, Api::Api& api,
-      Http::Context& http_context, Grpc::Context& grpc_context, Router::Context& router_context,
-      Server::Instance& server, Config::XdsManager& xds_manager) {
+      Server::Admin& admin, Api::Api& api, Http::Context& http_context, Grpc::Context& grpc_context,
+      Router::Context& router_context, Server::Instance& server, Config::XdsManager& xds_manager) {
     absl::Status creation_status = absl::OkStatus();
     auto cluster_manager = std::unique_ptr<TestClusterManagerImpl>{new TestClusterManagerImpl(
         bootstrap, factory, context, stats, tls, runtime, local_info, log_manager,
-        main_thread_dispatcher, admin, validation_context, api, http_context, grpc_context,
-        router_context, server, xds_manager, creation_status)};
+        main_thread_dispatcher, admin, api, http_context, grpc_context, router_context, server,
+        xds_manager, creation_status)};
     THROW_IF_NOT_OK(creation_status);
     THROW_IF_NOT_OK(cluster_manager->initialize(bootstrap));
     return cluster_manager;
@@ -212,18 +211,19 @@ public:
 protected:
   using ClusterManagerImpl::ClusterManagerImpl;
 
-  TestClusterManagerImpl(
-      const envoy::config::bootstrap::v3::Bootstrap& bootstrap, ClusterManagerFactory& factory,
-      Server::Configuration::CommonFactoryContext& context, Stats::Store& stats,
-      ThreadLocal::Instance& tls, Runtime::Loader& runtime, const LocalInfo::LocalInfo& local_info,
-      AccessLog::AccessLogManager& log_manager, Event::Dispatcher& main_thread_dispatcher,
-      Server::Admin& admin, ProtobufMessage::ValidationContext& validation_context, Api::Api& api,
-      Http::Context& http_context, Grpc::Context& grpc_context, Router::Context& router_context,
-      Server::Instance& server, Config::XdsManager& xds_manager, absl::Status& creation_status)
+  TestClusterManagerImpl(const envoy::config::bootstrap::v3::Bootstrap& bootstrap,
+                         ClusterManagerFactory& factory,
+                         Server::Configuration::CommonFactoryContext& context, Stats::Store& stats,
+                         ThreadLocal::Instance& tls, Runtime::Loader& runtime,
+                         const LocalInfo::LocalInfo& local_info,
+                         AccessLog::AccessLogManager& log_manager,
+                         Event::Dispatcher& main_thread_dispatcher, Server::Admin& admin,
+                         Api::Api& api, Http::Context& http_context, Grpc::Context& grpc_context,
+                         Router::Context& router_context, Server::Instance& server,
+                         Config::XdsManager& xds_manager, absl::Status& creation_status)
       : ClusterManagerImpl(bootstrap, factory, context, stats, tls, runtime, local_info,
-                           log_manager, main_thread_dispatcher, admin, validation_context, api,
-                           http_context, grpc_context, router_context, server, xds_manager,
-                           creation_status) {}
+                           log_manager, main_thread_dispatcher, admin, api, http_context,
+                           grpc_context, router_context, server, xds_manager, creation_status) {}
 };
 
 } // namespace Upstream
