+++ test/mocks/config/xds_manager.cc
@@ -7,6 +7,8 @@ using testing::Return;
 
 MockXdsManager::MockXdsManager() {
   ON_CALL(*this, initialize(_, _)).WillByDefault(Return(absl::OkStatus()));
+  ON_CALL(*this, initializeAdsConnections(_)).WillByDefault(Return(absl::OkStatus()));
+  ON_CALL(*this, adsMux()).WillByDefault(Return(ads_mux_));
   ON_CALL(*this, subscriptionFactory()).WillByDefault(ReturnRef(subscription_factory_));
 }
